Break up test_utils.rs: extract struct extensions
commit-id:d621942f
